backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
python-multipart
pyjwt[crypto]
passlib[bcrypt]
python-dotenv
pytest
httpx
requests>=2.31.0